backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.5
httpx==0.28.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiohttp==3.11.10 